Restores proper sqlite db name
When deploying to Heroku, the database gets switched to Postgresql.  I
had mistakenly thought the sqlite3 based name would screw that up, but
Heroku ignores the adaptor and db name paramters in database.yml. But my
changes caused the development and test databases to be in the project
directory instead of the db/ directory.

This commit restores the original style of the db name (and location)
and comments it for the next guy.